Implemented create and write atomic operations

+++ src/core/system/architecture/operations/create.py
@@ -4,21 +4,19 @@ import pathlib
 
 class Create:
 
-	def __init__(self, uri: pathlib.Path, data: bytes):
+	def __init__(self, uri: pathlib.Path):
 
 		self.uri = uri
-		self.data = data
+		self.existed = uri.exists()
 
-	def execute(self) -> bytes:
+	def execute(self) -> None:
 
 		self.uri.parent.mkdir(parents = True, exist_ok = True)
-		self.uri.write_bytes(self.data)
-
-		return self.data
+		self.uri.touch(exist_ok = True)
 
 	def rollback(self):
 
-		if self.uri.exists():
+		if self.uri.exists() and not self.existed:
 
 			self.uri.unlink()
